Refactor authentication module: update refresh token handling, improve error structure, and adjust import paths
